fixed tests

+++ DataParser/DataParser.cpp
@@ -30,14 +30,14 @@ void DataParser::handleInformationIncoming()
     QString string(data);
     if(!data.isNull() && string.at(0) == '#')
     {
-      for(int i=1;i<string.length()-1;i++){
-          if(string.at(i)<48||string.at(i)>57){
-              if(string.at(i)!=45)
+      for(int i = 1; i <string.length()-1; i++){
+          if(string.at(i) < 48||string.at(i) > 57){
+              if(string.at(i) != 45)
               return;
           }
       }
       int id = string.mid(1,2).toInt();
-      int value = string.mid(3).toInt();
+      double value = string.mid(3).toDouble();
       emit dataReceived(id, value);
       emit sendDebugMessage(data);
     }
